requirements.txt
aniso8601==9.0.1
anyio==3.6.2
beautifulsoup4==4.11.1
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
fastapi==0.89.1
Flask==2.2.2
Flask-RESTful==0.3.9
google-api-core==2.11.0
google-api-python-client==2.73.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
google-cloud==0.34.0
googleapis-common-protos==1.58.0
h11==0.14.0
httplib2==0.21.0
idna==3.4
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
mysql-connector==2.2.9
mysql-connector-python==8.0.32
numpy==1.24.1
oauthlib==3.2.2
pandas==1.5.3
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.4
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
starlette==0.22.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.14
uvicorn==0.20.0
waitress==2.1.2
Werkzeug==2.2.2
ytinspector @ git+https://github.com/DataSolveProblems/YTInspector.git@2c52c9cc94529b00dbbf597fda28d70ddc4cfc23
